% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_product_cube.R
\name{download_product_cube}
\alias{download_product_cube}
\title{Download StatsCan Product Cube}
\usage{
download_product_cube(productId)
}
\arguments{
\item{productId}{The Statistics Canada Product ID.}
}
\description{
This function allows you to download a product cube from
Statistics Canada.
}
\examples{
\dontrun{
download_product_cube('14100287')
}
}
\keyword{csv}
\keyword{cube,}
\keyword{product,}
\keyword{productId,}
