% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cansim_metadata.R
\name{get_cansim_metadata}
\alias{get_cansim_metadata}
\title{Download StatsCan Metadata from Legacy CASIM Table}
\usage{
get_cansim_metadata(cansimId)
}
\arguments{
\item{cansimId}{The Statistics Canada Legacy CANSIM ID.}
}
\description{
This function allows you to download product metadata from
Statistics Canada.
}
\examples{
get_cansim_metadata('2820087')

}
\keyword{CANSIM,}
\keyword{cansimId,}
\keyword{metadata}
