% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coordinate_data.R
\name{get_coordinate_data}
\alias{get_coordinate_data}
\title{Download StatsCan Coordinate Data from Product Cube}
\usage{
get_coordinate_data(productId, coordinateId, nperiods)
}
\arguments{
\item{productId}{The Statistics Canada Product ID.}

\item{coordinateId}{The Statistics Canada Coordinate ID (10 digits).}

\item{nperiods}{Number of time periods.}
}
\description{
This function allows you to download a coordinate series from
Statistics Canada.
}
\examples{
get_coordinate_data('14100287', '1.1.1.1.1.2.0.0.0.0', 3)

}
\keyword{coordinate}
\keyword{cube,}
\keyword{productId,}
