% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_range.R
\name{get_data_range}
\alias{get_data_range}
\title{Download StatsCan Vector Data by date range}
\usage{
get_data_range(productId = NA, coordinateId = NA, vectorId = NA,
  startDt, endDt)
}
\arguments{
\item{productId}{The Statistics Canada Product ID.}

\item{coordinateId}{The Statistics Canada Coordinate ID (10 digits).}

\item{vectorId}{The Statistics Canada Vector ID (10 digits).}
}
\description{
This function allows you to download a data range
}
\examples{
get_data_range(vectorId='2064888', startDt="2015-12-01T08:30", endDt="2018-03-31T19:00")

}
\keyword{coordinate}
\keyword{cube,}
\keyword{productId,}
