% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_product_metadata.R
\name{get_product_metadata}
\alias{get_product_metadata}
\title{Download StatsCan Metadata from Product Cube}
\usage{
get_product_metadata(productId)
}
\arguments{
\item{productId}{The Statistics Canada Product ID.}
}
\description{
This function allows you to download product metadata from
Statistics Canada.
}
\examples{
get_product_metadata('14100287')

}
\keyword{metadata}
\keyword{product,}
\keyword{productId,}
