% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cansim_product_mapping.R
\name{read_cansim_product_mapping}
\alias{read_cansim_product_mapping}
\title{A function to map legacy CANSIM ID to Product ID}
\usage{
read_cansim_product_mapping(cansimId = "all", productId = "all")
}
\arguments{
\item{cansimId}{The Statistics Canada CANSIM ID.}

\item{productId}{The Statistics Canada Product ID.}
}
\description{
This function allows you to either download the entire
Statistics Canada CANSIM ID to Product ID concordance or
map a single CANSIM ID to Product ID.
}
\examples{
read_cansim_product_mapping()
read_cansim_product_mapping(productId = '14100287')
read_cansim_product_mapping(cansimId = '2820087')

}
\keyword{cansimId,}
\keyword{concordance}
\keyword{csv,}
\keyword{productId,}
