% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statscanr.R
\docType{package}
\name{statscanr-package}
\alias{statscanr}
\alias{statscanr-package}
\title{statscanr: A Data Pipeline Package for Statistics Canada Web Data Services}
\description{
An R package for interacting with the Statistics Canada Web Data Services.
    The primary purpose of the package is to retrieve data from Statistics Canada for analytics or data pipelines.
}
\details{
The only R package for interacting with Statistics Canada Web Data Services.
}
\keyword{internal}
\keyword{statistics}
\keyword{statscan}
